We wrote a prettty important function
and we tested it

next time, and I'm going to write this in the commit log instead of in
the calendar invite, we need to use shift_and_read in some kind of
peek_integer() function. oh and we need to make a peek_integer()
function so we can read the block magic bytes.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>