turn symbol bitmaps into trees
First, we wrote a test for our canonical huffman code implementation
(the second example from the PDF). Amazingly, that passed.

Second, we wrote an algorithm to turn a `SymbolBitmap` (the output of
our canonical huffman code) into a huffman tree as previously
implemented. It does this by walking the paths and dynamically creating
branch and leaf nodes as needed. We also wrote some basic tests for this
functionality.