Add another test on burrows_wheeler
In our quest to figure out why we have an output that is all 1's for our
test case, we were suspicious of our burrows wheeler implementation. We
were able to add a test inspired by Wikipedia's example, which convinced
us that our implementation is working correctly, or at least in
agreement with how Wikipedia describes the algorithm.

We also cleaned up some comments and added new ones.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>