Begin, but do not complete fixing the symbol stack
We had missed that the symbol stack was encoded in the file format,
rather than starting off as 0..255, which is what our impl assumed. In
this commit we began to fix that issue, but ran out of time, and
realized that we had misinterpreted the length of l2 in the SymbolMap.

We left a TODONEXT with what we think needs to be done to fix that
issue, which we can take care of next time.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>