Rename the CI build/test 1.89 to msrv
msrv means "minimum supported rust version". This allows us to configure
CI to require a test to pass called msrv, rather than requiring a test
to pass called 1.89. This means when we bump msrv in the future, we
don't have to adjust CI rules for branch protection, because it will
still be named msrv.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>